requirements.txt
requests~=2.24.0
pyyaml
